refactor: move `getCredentialType()` to lib/utils dir
