+++ src/lib/utils/getCredentialType.ts
@@ -1,0 +1,7 @@
+export function getCredentialType(parsedCredential: any): string {
+	return (
+		parsedCredential?.metadata?.credential?.vct ??
+		parsedCredential?.metadata?.credential?.doctype ??
+		''
+	);
+}
